rename misnomer CodeBlocks -> ChatMessageContent
This component shows the entire chat message content, not just code blocks.
