+++ vscode/webviews/chat/ChatMessageContent.module.css
